% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{A_func}
\alias{A_func}
\title{Sum of Squares (helper function)}
\usage{
A_func(matrix)
}
\arguments{
\item{matrix}{data matrix}
}
\description{
Sum of Squares (helper function)
}
\keyword{internal}
