% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ahat1_func}
\alias{ahat1_func}
\title{Estimator for expansion term in Frobenius Norm Schott 2007 (helper funciton)}
\usage{
ahat1_func(p, sample_cov)
}
\arguments{
\item{p}{dimension}

\item{sample_cov}{sample covariance matrix}
}
\value{

}
\description{
Estimator for expansion term in Frobenius Norm Schott 2007 (helper funciton)
}

