% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ahat2_func}
\alias{ahat2_func}
\title{Estimator for expansion term in Frobenius Norm Schott 2007 (helper funciton)}
\usage{
ahat2_func(n1, n2, p, overall_cov)
}
\arguments{
\item{p}{dimension}

\item{n}{sample size}

\item{sample_cov}{sample covariance matrix}
}
\description{
Estimator for expansion term in Frobenius Norm Schott 2007 (helper funciton)
}
\keyword{internal}

