% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_functions.R
\name{ahat2i_func}
\alias{ahat2i_func}
\title{Estimator for expansion term in Frobenius Norm Schott 2007 (helper funciton)}
\usage{
ahat2i_func(n, p, sample_covs)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{sample_covs}{sample covariance matrix}
}
\description{
Estimator for expansion term in Frobenius Norm Schott 2007 (helper funciton)
}
\keyword{internal}

