% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ahat2iSrivastava2014_func}
\alias{ahat2iSrivastava2014_func}
\title{Estimator for frobenius norm expansion term Srivastava 2014 (helper function)}
\usage{
ahat2iSrivastava2014_func(n, p, D, A)
}
\arguments{
\item{n}{sample size for groups}

\item{A}{sum of squares}

\item{sample_covs}{covariance matrices for groups}

\item{see}{function}

\item{overall_n}{overall sample size}
}
\description{
Estimator for frobenius norm expansion term Srivastava 2014 (helper function)
}
\keyword{internal}

