% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_functions.R
\name{ahat2Srivastava2014_func}
\alias{ahat2Srivastava2014_func}
\title{Estimator for frobenius norm expansion term Srivastava 2014 (helper function)}
\usage{
ahat2Srivastava2014_func(ahat2i, n)
}
\arguments{
\item{ahat2i}{see function}

\item{n}{sample size group 1}
}
\description{
Estimator for frobenius norm expansion term Srivastava 2014 (helper function)
}
\keyword{internal}
