% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_functions.R
\name{ahat4_func}
\alias{ahat4_func}
\title{Estimator for frobenius norm expansion term Srivastava 2007}
\usage{
ahat4_func(A, p, ns, ahat2, ahat1)
}
\arguments{
\item{A}{sum of squares for group 1}

\item{p}{dimension}

\item{ns}{sample size for group 1}

\item{ahat2}{see function}

\item{ahat1}{see function}
}
\description{
Estimator for frobenius norm expansion term Srivastava 2007
}
\keyword{internal}

