% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_functions.R
\name{ahatStar4_func}
\alias{ahatStar4_func}
\title{Estimator for expansion term in Frobenius Norm Chaipitak 2013 (helper funciton)}
\usage{
ahatStar4_func(tau, p, sample_cov, n1, n2)
}
\arguments{
\item{tau}{see function}

\item{p}{dimension}

\item{sample_cov}{sample covariance matrix}

\item{n1}{sample size for group 1}

\item{n2}{sample size for group 2}
}
\description{
Estimator for expansion term in Frobenius Norm Chaipitak 2013 (helper funciton)
}
\keyword{internal}

