% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_functions.R
\name{bhat_func}
\alias{bhat_func}
\title{b hat used in method introduced by Chaipitak 2013 (helper function)}
\usage{
bhat_func(ahat21, ahat22)
}
\arguments{
\item{ahat21}{see function}

\item{ahat22}{see function}
}
\description{
b hat used in method introduced by Chaipitak 2013 (helper function)
}
\keyword{internal}
