% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxesM_test.R
\name{BoxesM}
\alias{BoxesM}
\title{Hidden Test}
\usage{
BoxesM(ns, n_overall, sample_covs, overall_cov)
}
\arguments{
\item{ns}{ns}

\item{n_overall}{n overall}

\item{sample_covs}{sample covs}

\item{overall_cov}{overall cov}
}
\description{
Hidden Test
}

