% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxesM_test.R
\name{BoxesM_test}
\alias{BoxesM_test}
\title{Test of Equality of Covariances given by Chaipitak 2013}
\usage{
BoxesM_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\value{
Test statistic for Chaipitak 2013
}
\description{
Test of Equality of Covariances given by Chaipitak 2013
}
\examples{
BoxesM_test(mcSamples(rep(0, 10), diag(1, 10), 100, 3), group = population)

}

