% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxesM_test.R
\name{BoxsM_test}
\alias{BoxsM_test}
\title{Test of Equality of Covariances given by Box's M}
\usage{
BoxsM_test(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test statistic for Chaipitak 2013
}
\description{
Test of Equality of Covariances given by Box's M
}
\examples{
BoxsM_test(iris, group = Species)

}
