% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bp_function.R
\name{Bp_}
\alias{Bp_}
\title{Base for Brewer Kernel Quantile Estimator}
\usage{
Bp_(n, i, p, X)
}
\arguments{
\item{n}{sample size}

\item{i}{index}

\item{p}{quantile}

\item{X}{ith data point}
}
\description{
Base for Brewer Kernel Quantile Estimator
}
\keyword{internal}

