% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bp_function.R
\name{Bp_func}
\alias{Bp_func}
\title{Brewer Kernel Quantile Estimator}
\usage{
Bp_func(data, p)
}
\arguments{
\item{data}{vector of values}

\item{p}{quantile}
}
\description{
Brewer Kernel Quantile Estimator
}
\examples{
Bp_func(rnorm(10000), .95)
}
\keyword{internal}

