% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{c0_func}
\alias{c0_func}
\title{Sample size expansion term Srivastava 2007 (helper function)}
\usage{
c0_func(n)
}
\arguments{
\item{n}{sample size}
}
\description{
Sample size expansion term Srivastava 2007 (helper function)
}
\keyword{internal}

