% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{c3_func}
\alias{c3_func}
\title{Sample size expansion term Srivastava 2007 (helper function)}
\usage{
c3_func(n)
}
\arguments{
\item{n}{sample size}
}
\value{

}
\description{
Sample size expansion term Srivastava 2007 (helper function)
}

