% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chaipitak2013_test.R
\name{Chaipitak2013 Test}
\alias{Chaipitak2013 Test}
\title{Test of Equality of Covariances given by Chaipitak and Chongcharoen 2013}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{group}{group or population variable}

\item{...}{other options passed to functions}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Chaipitak and Chongcharoen 2013
}
\details{
\deqn{T_{Sc} = \sum \limits^{k}_{i < j} \frac{ \left( \hat{a}_{2i_{Sc}} + \hat{a}_{2j_{Sc}} - \frac{2}{p}tr \left( S_i S_j \right) \right) ^ 2}{\theta_{Sc}}}
}
\section{Usage}{

\preformatted{Chaipitak2013_test(x, ...)

## S3 method for class 'data.frame'
Chaipitak2013_test(x, group, ...)

## S3 method for class 'grouped_df'
Chaipitak2013_test(x, ...)

## S3 method for class 'resample'
Chaipitak2013_test(x, ...)

## S3 method for class 'matrix'
Chaipitak2013_test(...)}
}
\examples{
Chaipitak2013_test(iris, group = Species)
}
\references{
Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality
of two covariance matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270–277.
}

