% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chaipitak2013_test.R
\name{Chaipitak2013}
\alias{Chaipitak2013}
\title{Hidden Test}
\usage{
Chaipitak2013(ahat2i, deltahat2)
}
\arguments{
\item{ahat2i}{a hat squared i}

\item{deltahat2}{delta hat squared}
}
\description{
Hidden Test
}

