% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chaipitak2013_test.R
\name{Chaipitak2013_test}
\alias{Chaipitak2013_test}
\alias{Chaipitak2013_test.data.frame}
\alias{Chaipitak2013_test.grouped_df}
\alias{Chaipitak2013_test.matrix}
\title{Test of Equality of Covariances given by Chaipitak 2013}
\usage{
Chaipitak2013_test(x, ...)

\method{Chaipitak2013_test}{data.frame}(x, group, ...)

\method{Chaipitak2013_test}{grouped_df}(x, ...)

\method{Chaipitak2013_test}{matrix}(...)
}
\arguments{
\item{x}{tidy data frame}

\item{...}{other}

\item{group}{group}
}
\value{
Test statistic for Chaipitak 2013
}
\description{
Test of Equality of Covariances given by Chaipitak 2013
}
\examples{
Chaipitak2013_test(iris, group = Species)

}

