% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChaipitakSimData.R
\name{ChaipitakSimData}
\alias{ChaipitakSimData}
\title{Simulated Data for Chaipitak and Chongaroen 2013}
\usage{
ChaipitakSimData(dim, n, covStructure = "unstructured", ...,
  DataMatrix = FALSE)
}
\arguments{
\item{dim}{dimensions for covariance matrix}

\item{n}{number of samples}

\item{covStructure}{covariance structure}

\item{...}{other options passed to functions}

\item{DataMatrix}{whether to output a data matrix or covariance matrix}
}
\value{
Data either covariance matrix or data matrix
}
\description{
Simulated Data for Chaipitak and Chongaroen 2013
}

