% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_functions.R
\name{ci_func}
\alias{ci_func}
\title{Sample size expansion term Srivastava 2007 (helper function)}
\usage{
ci_func(n, p)
}
\arguments{
\item{n}{sample size}
}
\description{
Sample size expansion term Srivastava 2007 (helper function)
}
\keyword{internal}

