% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_maker.R
\name{cov_maker}
\alias{cov_maker}
\title{Covariance Maker}
\usage{
cov_maker(keepers, offs, losers)
}
\arguments{
\item{losers}{}
}
\description{
Covariance Maker
}
\examples{
cov_maker(keepers = list(c(20, 1, rep(5, 8))),
                    offs = list(.1, nrow = 10, ncol = 90),
                    losers = list(c(1, rep(.1, 89))))
}

