% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covTestR-package.R
\docType{package}
\name{covTestR-package}
\alias{covTestR-package}
\title{Covariance Matrix Testing Functions}
\description{
Testing functions for Covariance Matrices. These tests 
include high-dimension homogeneity of covariance matrix testing described 
by Schott (2007) \href{http://doi.org/10.1016/j.csda.2007.03.004}{10.1016/j.csda.2007.03.004} 
and high-dimensional one-sample tests of covariance matrix structure 
described by Fisher, et al. (2010) \href{http://doi.org/10.1016/j.jmva.2010.07.004}{10.1016/j.jmva.2010.07.004}. 
Covariance matrix tests use C++ to speed performance and allow larger data sets.
}
