% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_sim.R
\name{crit_data}
\alias{crit_data}
\title{Title}
\usage{
crit_data(type, dimensions, samples, difference, populations, replications,
  reductionMethods, directory)
}
\arguments{
\item{directory}{}
}
\description{
Title
}
\examples{
crit_data(type = "identity",
                    dimensions = c(20, 25, 30, 35, 40, 45, 50),
                    samples = c(15),
                    difference = c(0, 1, 2, 3, 4, 5),
                    populations = 2,
                    replications = 10,
                    reductionMethods = c("DataConcatScatter", "DataConcatScatterBlock",
                    "SConcat", "SDiff"),
                    directory = "~/Documents/R/Dissertation/data")

}

