% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_sim3.R
\name{crit_data3}
\alias{crit_data3}
\title{Title}
\usage{
crit_data3(type, dimensions, samples, difference, populations, replications,
  reductionMethods, directory)
}
\arguments{
\item{directory}{}
}
\description{
Title
}
\examples{
crit_data3(type = "identity",
                    dimensions = c(100, 125, 150, 175, 200),
                    samples = c(15),
                    difference = c(0, 1, 2, 3, 4, 5),
                    populations = 3,
                    replications = 1000,
                    reductionMethods = c("DataConcatScatter", "DataConcatScatterBlock",
                    "SConcat", "SDiff"),
                    directory = "~/Documents/R/Dissertation/data")

}

