% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{critical_value_sim_}
\alias{critical_value_sim_}
\title{Sim Function to get data for Critical Value Data}
\usage{
critical_value_sim_(data, quantile, ..., dots)
}
\arguments{
\item{data}{the data}

\item{...}{other stuff}
}
\description{
Sim Function to get data for Critical Value Data
}
\examples{
critical_value_sim_(mcSamples(c(0,0,0), diag(1, 3), 10, 2))
}
\keyword{internal}

