% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dataDftoMatrix}
\alias{dataDftoMatrix}
\title{Turn no a Tidy data frame into data matrix (helper function)}
\usage{
dataDftoMatrix(data, group, method, ..., .dots)
}
\arguments{
\item{data}{not a tidy dataframe}

\item{group}{group}

\item{method}{method}

\item{...}{other options}

\item{.dots}{other options}
}
\description{
Turn no a Tidy data frame into data matrix (helper function)
}
\keyword{internal}

