% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{deltahat2_func}
\alias{deltahat2_func}
\title{delta hat squared for Chaiptak 2013 (helper function)}
\usage{
deltahat2_func(ahatstar4, p, ahat2, n1, n2)
}
\arguments{
\item{ahatstar4}{see function}

\item{p}{dimension}

\item{ahat2}{see function}

\item{n1}{sample size for group 1}

\item{n2}{sample size for group 2}
}
\description{
delta hat squared for Chaiptak 2013 (helper function)
}
\keyword{internal}

