% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{Di_func}
\alias{Di_func}
\title{D diagonal matrix for Srivastava 2014 (helper funciton)}
\usage{
Di_func(matrix)
}
\arguments{
\item{matrix}{data matrix}
}
\description{
D diagonal matrix for Srivastava 2014 (helper funciton)
}
\keyword{internal}

