% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_trace.R
\name{diff_trace}
\alias{diff_trace}
\title{Difference in Trace of Covariance Matrices}
\usage{
diff_trace(data, ...)
}
\arguments{
\item{data}{Data}

\item{...}{other}
}
\value{
Difference of Traces for Covariance Matrices
}
\description{
Difference in Trace of Covariance Matrices
}
\examples{
diff_trace(mcSamples(c(0,0,0), diag(1, 3), 10, 2), group = population)

}

