% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_data.frame.R
\name{dot_data.frame}
\alias{dot_data.frame}
\title{Change data to matrix helper}
\usage{
dot_data.frame(test, data, group, ..., tidy = FALSE)
}
\arguments{
\item{tidy}{}
}
\description{
Change data to matrix helper
}

