% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_functions.R
\name{etahat2i_func}
\alias{etahat2i_func}
\title{Estimator for eta squared Srivastave 2007 (helper function)}
\usage{
etahat2i_func(n, p, ahat4, ahat2)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{ahat4}{see function}

\item{ahat2}{see function}
}
\description{
Estimator for eta squared Srivastave 2007 (helper function)
}
\keyword{internal}
