% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher2012.R
\name{Fisher2012}
\alias{Fisher2012}
\title{Test of Equality of Covariances given by Chaipitak and Chongcharoen 2013}
\usage{
Fisher2012(x, Sigma = "identity", ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Chaipitak and Chongcharoen 2013
}
\examples{
Fisher2012(as.matrix(iris[1:50, 1:4]))

}
\references{
Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality of two covariance
matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270-277.
}
