% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{gammahati_func}
\alias{gammahati_func}
\title{Estimator for Gamma Srivastava and Yanagihara 2010 (helper function)}
\usage{
gammahati_func(ahat2i, ahat1i)
}
\arguments{
\item{ahat2i}{see function}

\item{ahat1i}{see function}
}
\description{
Estimator for Gamma Srivastava and Yanagihara 2010 (helper function)
}
\keyword{internal}
