% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ishii2016.R
\name{Ishii2016}
\alias{Ishii2016}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Ishii2016(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test Statistic for
}
\description{
Test of Equality of Covariances given by Schott 2007
}
\examples{
irisSpecies <- unique(iris$Species)

iris_ls <- lapply(irisSpecies, 
    function(x){as.matrix(iris[iris$Species == x, 1:4])}
                 )
                 
names(iris_ls) <- irisSpecies

Chaipitak2013(iris_ls)
}
\references{
Ishii, A., Yata, K., and Aoshima, M. (2016). Asymptotic properties of the first pricipal
component and equality tests of covariance matrices in high-dimesion, low-sample-size context. Journal
of Statistical Planning and Inference, 170:186-199.
}
