% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksi_functions.R
\name{ksihat2i_func}
\alias{ksihat2i_func}
\title{Estimator for Ksi Srivastava and Yanagihara 2010 (helper function)}
\usage{
ksihat2i_func(n, p, ahat1, ahat2, ahat3, ahat4)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{ahat1}{see function}

\item{ahat2}{see function}

\item{ahat3}{see function}

\item{ahat4}{see function}
}
\description{
Estimator for Ksi Srivastava and Yanagihara 2010 (helper function)
}
\keyword{internal}

