% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LedoitWolf2002_test.R
\name{LedoitWolf2002_test}
\alias{LedoitWolf2002_test}
\title{Test of Equality of Covariances given by Chaipitak and Chongcharoen 2013}
\usage{
LedoitWolf2002_test(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}

\item{covMat}{Covariance Matrix}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Chaipitak and Chongcharoen 2013
}
\examples{
LedoitWolf2002W_test(iris[1:50, 1:3])

}
\references{
Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality of two covariance
matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270-277.
}
