% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcSamples.R
\name{mcSamples}
\alias{mcSamples}
\title{Simulate Monte Carlo Samples from a Multivariate Normal Distribution}
\usage{
mcSamples(meanVec, covMat, samples, pops, ...)
}
\arguments{
\item{meanVec}{Numeric vector of means for the variables.}

\item{covMat}{Positive-definite symmetric matrix of the variables.}

\item{...}{Other variables used in mvnorm function in the mass package.}

\item{maxN}{Numeric scalar of the maximum of the sample sizes in the simulation.}
}
\value{
Matrix of maxN multivariate normal samples.
}
\description{
Simulate Monte Carlo Samples from a Multivariate Normal Distribution
}
\examples{
mcSamples(c(0,0,0), diag(1, 3), 10, 2)
}

