% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{overall_cov_func}
\alias{overall_cov_func}
\title{Overall Covariance Matirx (helper function)}
\usage{
overall_cov_func(A_ls, ns)
}
\arguments{
\item{A_ls}{sum of squares for group 1}

\item{ns}{sample size for group 1}
}
\description{
Overall Covariance Matirx (helper function)
}
\keyword{internal}
