% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schott2007_test.R
\name{Schott2007}
\alias{Schott2007}
\title{Hidden Test}
\usage{
Schott2007(ns, p, ahat2, ahat2i, sample_covs)
}
\arguments{
\item{ns}{ns}

\item{p}{p}

\item{ahat2}{a hat squared}

\item{ahat2i}{a hat squared i}

\item{sample_covs}{sample covs}
}
\description{
Hidden Test
}

