% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schott2007sample.R
\name{Schott2007sample}
\alias{Schott2007sample}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Schott2007sample(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Schott 2007
}
\examples{
Schott2007(iris, group = Species)

}
\references{
Schott, J. (2007). A test for the equality of covariance matrices when the dimension
is large relative to the sample sizes. Computational Statistics & Data Analysis, 51(12):6535-6542.
}
