% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songStructure.R
\name{songStructure}
\alias{songStructure}
\title{Song Conical Variates For Structure of Covariance Tests}
\usage{
songStructure(x, Sigma)
}
\arguments{
\item{Sigma}{}
}
\description{
Song Conical Variates For Structure of Covariance Tests
}
