% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2005.R
\name{Srivastava2005}
\alias{Srivastava2005}
\title{Test of Equality of Covariances given by Chaipitak and Chongcharoen 2013}
\usage{
Srivastava2005(x, Sigma = "identity", ...)
}
\arguments{
\item{x}{data}

\item{Sigma}{Population covariance matrix}

\item{...}{other options passed to covTest method}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Chaipitak and Chongcharoen 2013
}
\examples{
Srivastava2005(as.matrix(iris[1:50, 1:4]))

}
\references{
Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality of two covariance
matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270-277.
}
