% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2007_test.R
\name{Srivastava2007}
\alias{Srivastava2007}
\title{Hidden Test}
\usage{
Srivastava2007(ahat2i, etahat2i)
}
\arguments{
\item{ahat2i}{a hat squared i}

\item{etahat2i}{eta hat squared i}
}
\description{
Hidden Test
}

