% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2007_test.R
\name{Srivastava2007_test}
\alias{Srivastava2007_test}
\title{Test of Equality of Covariances given by Srivastava 2007}
\usage{
Srivastava2007_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\value{
Test statistic for Srivastava 2007
}
\description{
Test of Equality of Covariances given by Srivastava 2007
}
\examples{
Srivastava2007_test(mcSamples(rep(0, 100), diag(1, 100), 10, 3), group = population)

}

