% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2014_test.R
\name{Srivastava2014}
\alias{Srivastava2014}
\title{Hidden Test}
\usage{
Srivastava2014(ns, p, ahat2, ahat2i, sample_covs)
}
\arguments{
\item{ns}{ns}

\item{p}{p}

\item{ahat2}{a hat squared}

\item{ahat2i}{a hat squared i}

\item{sample_covs}{sample covs}
}
\description{
Hidden Test
}

