% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2014_test.R
\name{Srivastava2014_test}
\alias{Srivastava2014_test}
\alias{Srivastava2014_test.data.frame}
\alias{Srivastava2014_test.grouped_df}
\alias{Srivastava2014_test.matrix}
\title{Test of Equality of Covariances given by Srivastava 2014}
\usage{
Srivastava2014_test(x, ...)

\method{Srivastava2014_test}{data.frame}(x, group, ...)

\method{Srivastava2014_test}{grouped_df}(x, ...)

\method{Srivastava2014_test}{matrix}(...)
}
\arguments{
\item{x}{tidy data frame}

\item{...}{other}

\item{group}{group}
}
\value{
Test Statistic for Srivastava 2014
}
\description{
Test of Equality of Covariances given by Srivastava 2014
}
\examples{
Srivastava2014_test(iris, group = Species)

}

