% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrivastavaYanagihara2010.R
\name{SrivastavaYanagihara2010}
\alias{SrivastavaYanagihara2010}
\title{Test of Equality of Covariances given by Srivastava and Yanagihara 2010}
\usage{
SrivastavaYanagihara2010(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test statistic for Srivastava and Yanagihara 2010
}
\description{
Test of Equality of Covariances given by Srivastava and Yanagihara 2010
}
\examples{
irisSpecies <- unique(iris$Species)

iris_ls <- lapply(irisSpecies, 
    function(x){as.matrix(iris[iris$Species == x, 1:4])}
                 )
                 
names(iris_ls) <- irisSpecies

Chaipitak2013(iris_ls)
}
\references{
Srivastava, M. and Yanagihara, H. (2010). Testing the equality of several covariance matrices with
fewer observation that the dimension. Journal of Multivariate Analysis, 101(6):1319-1329.
}
