% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrivastavaYanagihara2010_test.R
\name{SrivastavaYanagihara2010_test}
\alias{SrivastavaYanagihara2010_test}
\alias{SrivastavaYanagihara2010_test.data.frame}
\alias{SrivastavaYanagihara2010_test.grouped_df}
\alias{SrivastavaYanagihara2010_test.matrix}
\title{Test of Equality of Covariances given by Srivastava and Yanagihara 2010}
\usage{
SrivastavaYanagihara2010_test(x, ...)

\method{SrivastavaYanagihara2010_test}{data.frame}(x, group, ...)

\method{SrivastavaYanagihara2010_test}{grouped_df}(x, ...)

\method{SrivastavaYanagihara2010_test}{matrix}(...)
}
\arguments{
\item{x}{tidy data frame}

\item{...}{other}

\item{group}{group}
}
\value{
Test statistic for Srivastava and Yanagihara 2010
}
\description{
Test of Equality of Covariances given by Srivastava and Yanagihara 2010
}
\examples{
SrivastavaYanagihara2010_test(iris, group = Species)

}

