% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{tau_func}
\alias{tau_func}
\title{Tau for Chaipitak 2013 (helper function)}
\usage{
tau_func(n1, n2)
}
\arguments{
\item{n}{sample size}
}
\description{
Tau for Chaipitak 2013 (helper function)
}
\keyword{internal}

