% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_sim.R
\name{tests}
\alias{tests}
\title{Title}
\usage{
tests(directory, save, test_funcs, dimensions)
}
\arguments{
\item{type}{}
}
\description{
Title
}
\examples{
tests(directory = "~/Documents/R/Dissertation/data",
                test_funcs = c("Schott2007_test", "Srivastava2007_test"),
                dimensions = seq(20, 5))
}

