% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_functions.R
\name{ahat2iSrivastava2014_func}
\alias{ahat2iSrivastava2014_func}
\title{Estimator for frobenius norm expansion term Srivastava 2014 (helper function)}
\usage{
ahat2iSrivastava2014_func(n, p, D, A)
}
\arguments{
\item{n}{sample size for groups}

\item{p}{function}

\item{D}{overall sample size}

\item{A}{sum of squares}
}
\description{
Estimator for frobenius norm expansion term Srivastava 2014 (helper function)
}
\keyword{internal}
