% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ahat2Srivastava2014_func}
\alias{ahat2Srivastava2014_func}
\title{Estimator for frobenius norm expansion term Srivastava 2014 (helper function)}
\usage{
ahat2Srivastava2014_func(ahat21, ahat22, n1, n2)
}
\arguments{
\item{ahat21}{see function}

\item{ahat22}{see funciton}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}
}
\description{
Estimator for frobenius norm expansion term Srivastava 2014 (helper function)
}
\keyword{internal}

