% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_functions.R
\name{ahat3_func}
\alias{ahat3_func}
\title{Estimator for Frobenius norm expansion term (helper function)}
\usage{
ahat3_func(A, p, ns, ahat2, ahat1)
}
\arguments{
\item{p}{dimension}

\item{ns}{sample size for group 1}

\item{ahat2}{see function}

\item{ahat1}{see function}

\item{A1}{sum of squares for group 1}
}
\description{
Estimator for Frobenius norm expansion term (helper function)
}
\keyword{internal}
