% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ahat3_func}
\alias{ahat3_func}
\title{Estimator for Frobenius norm expansion term (helper function)}
\usage{
ahat3_func(A1, A2, p, n1, n2, ahat2, ahat1)
}
\arguments{
\item{A1}{sum of squares for group 1}

\item{A2}{sum of squares for group 2}

\item{p}{dimension}

\item{n1}{sample size for group 1}

\item{n2}{sample size for group 2}

\item{ahat2}{see function}

\item{ahat1}{see function}
}
\description{
Estimator for Frobenius norm expansion term (helper function)
}
\keyword{internal}

