% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_functions.R
\name{ahat4_func}
\alias{ahat4_func}
\title{Estimator for frobenius norm expansion term Srivastava 2007}
\usage{
ahat4_func(A, p, ns, ahat2, ahat1)
}
\arguments{
\item{p}{dimension}

\item{ahat2}{see function}

\item{ahat1}{see function}

\item{A1}{sum of squares for group 1}

\item{A2}{sum of squares for group 2}

\item{n1}{sample size for group 1}

\item{n2}{sample size for group 2}
}
\description{
Estimator for frobenius norm expansion term Srivastava 2007
}
\keyword{internal}

