% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ahmad2015.R
\name{Ahmad2015}
\alias{Ahmad2015}
\title{Test of Equality of Covariances given by Chaipitak and Chongcharoen 2013}
\usage{
Ahmad2015(x, Sigma = "identity", ...)
}
\arguments{
\item{x}{data}

\item{Sigma}{Population covariance matrix}

\item{...}{other options passed to covTest method}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Chaipitak and Chongcharoen 2013
}
\examples{
structureCovariances(iris[,1:4], covTest = Ahmad2015)

}
\references{
Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality of two covariance
matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270-277.
}
