% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ahmad2017.R
\name{Ahmad2017}
\alias{Ahmad2017}
\title{Test of Equality of Covariances given by Chaipitak and Chongcharoen 2013}
\usage{
Ahmad2017(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Chaipitak and Chongcharoen 2013
}
\examples{
irisSpecies <- unique(iris$Species)

iris_ls <- lapply(irisSpecies, 
    function(x){as.matrix(iris[iris$Species == x, 1:4])}
                 )
                 
names(iris_ls) <- irisSpecies

Chaipitak2013(iris_ls)
}
\references{
Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality of two covariance
matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270-277.
}
