% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxesM.R
\name{BoxesM}
\alias{BoxesM}
\title{Test of Equality of Covariances given by Box's M}
\usage{
BoxesM(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test statistic for Chaipitak 2013
}
\description{
Test of Equality of Covariances given by Box's M
}
\examples{
irisSpecies <- unique(iris$Species)

iris_ls <- lapply(irisSpecies, 
    function(x){as.matrix(iris[iris$Species == x, 1:4])}
                 )
                 
names(iris_ls) <- irisSpecies

Chaipitak2013(iris_ls)
}
