% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxesM_test.R
\name{BoxesM_test}
\alias{BoxesM_test}
\alias{BoxesM_test.data.frame}
\alias{BoxesM_test.grouped_df}
\alias{BoxesM_test.matrix}
\title{Test of Equality of Covariances given by Chaipitak 2013}
\usage{
BoxesM_test(x, ...)

\method{BoxesM_test}{data.frame}(x, group, ...)

\method{BoxesM_test}{grouped_df}(x, ...)

\method{BoxesM_test}{matrix}(...)
}
\arguments{
\item{x}{tidy data frame}

\item{...}{other}

\item{group}{group}
}
\value{
Test statistic for Chaipitak 2013
}
\description{
Test of Equality of Covariances given by Chaipitak 2013
}
\examples{
BoxesM_test(iris, group = Species)

}

