% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{Chaipitak2013_test}
\alias{Chaipitak2013_test}
\title{Test of Equality of Covariances given by Chaipitak 2013}
\usage{
Chaipitak2013_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\description{
Test of Equality of Covariances given by Chaipitak 2013
}

