% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chaipitak2013_test.R
\name{Chaipitak2013_test}
\alias{Chaipitak2013_test}
\title{Test of Equality of Covariances given by Chaipitak 2013}
\usage{
Chaipitak2013_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\value{
Test statistic for Chaipitak 2013
}
\description{
Test of Equality of Covariances given by Chaipitak 2013
}
\examples{
Chaipitak2013_test(mcSamples(rep(0, 100), diag(1, 100), 10, 3), group = population)

}

