% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chaipitak2013_test.R
\name{Chaipitak2013_test}
\alias{Chaipitak2013_test}
\title{Test of Equality of Covariances given by Chaipitak and Chongcharoen 2013}
\usage{
Chaipitak2013_test(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}

\item{group}{group or population variable}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Chaipitak and Chongcharoen 2013
}
\examples{
Chaipitak2013_test(iris, group = Species)

}
\references{
Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality of two covariance
matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270-277.
}
