% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{critical_value_sim}
\alias{critical_value_sim}
\title{Title}
\usage{
critical_value_sim(data)
}
\arguments{
\item{data}{}
}
\description{
Title
}
\examples{
critical_value_sim(mcSamples(c(0,0,0), diag(1, 3), 10, 2))
}

