% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_value_sim.R
\name{critical_value_sim}
\alias{critical_value_sim}
\title{Critical Value Simulation}
\usage{
critical_value_sim(sim_size, sim_data_func, quantile, ...)
}
\arguments{
\item{sim_data_func}{}
}
\description{
Critical Value Simulation
}
\keyword{internal}

