% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_value_sim.R
\name{critical_value_sim}
\alias{critical_value_sim}
\title{Sim Function to get data for Critical Value Data}
\usage{
critical_value_sim(params, replications, quantile, test)
}
\arguments{
\item{data}{the data}

\item{...}{other stuff}
}
\description{
Sim Function to get data for Critical Value Data
}
\examples{
critical_value_sim(list(meanVec = rep(0, 100),
                                  covMat = diag(1, 100),
                                  samples = 10,
                                  pops = 2),
                             100, .95, Chaipitak2013_test)
}
\keyword{internal}

