% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dataDftoMatrix}
\alias{dataDftoMatrix}
\title{Turn Tidy data frame into data matrix (helper function)}
\usage{
dataDftoMatrix(data)
}
\arguments{
\item{data}{tidy dataframe}
}
\description{
Turn Tidy data frame into data matrix (helper function)
}
\keyword{internal}

