% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_functions.R
\name{deltahat2_func}
\alias{deltahat2_func}
\title{delta hat squared for Chaiptak 2013 (helper function)}
\usage{
deltahat2_func(ahatstar4, p, ahat2, n)
}
\arguments{
\item{ahatstar4}{see function}

\item{p}{dimension}

\item{ahat2}{see function}

\item{n}{sample size}
}
\description{
delta hat squared for Chaiptak 2013 (helper function)
}
\keyword{internal}

