% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ishii2016_test.R
\name{Ishii2016}
\alias{Ishii2016}
\title{Hidden Test}
\usage{
Ishii2016(lambdatildes, htilde, ki)
}
\arguments{
\item{lambdatildes}{lambda tildes}

\item{htilde}{h tilde}

\item{ki}{ki}
}
\description{
Hidden Test
}
