% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ishii2016_test.R
\name{Ishii2016_test}
\alias{Ishii2016_test}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Ishii2016_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\value{
Test Statistic for Schott 2007
}
\description{
Test of Equality of Covariances given by Schott 2007
}
\examples{
Ishii2016_test(mcSamples(rep(0, 100), diag(1, 100), 10, 3), group = population)

}

