% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ishii2016_test.R
\name{Ishii2016_test}
\alias{Ishii2016_test}
\alias{Ishii2016_test.data.frame}
\alias{Ishii2016_test.grouped_df}
\alias{Ishii2016_test.matrix}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Ishii2016_test(x, ...)

\method{Ishii2016_test}{data.frame}(x, group, ...)

\method{Ishii2016_test}{grouped_df}(x, ...)

\method{Ishii2016_test}{matrix}(...)
}
\arguments{
\item{x}{data}

\item{...}{other}

\item{group}{gropuing variable}
}
\value{
Test Statistic for
}
\description{
Test of Equality of Covariances given by Schott 2007
}
\examples{
Ishii2016_test(iris, group = Species)
}

