% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ishii2016_test.R
\name{Ishii2016_test}
\alias{Ishii2016_test}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Ishii2016_test(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test Statistic for
}
\description{
Test of Equality of Covariances given by Schott 2007
}
\examples{
Ishii2016_test(iris, group = Species)
}
\references{
Ishii, A., Yata, K., and Aoshima, M. (2016). Asymptotic properties of the first pricipal
component and equality tests of covariance matrices in high-dimesion, low-sample-size context. Journal
of Statistical Planning and Inference, 170:186-199.
}
