% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcSamples.R
\name{mcSamples}
\alias{mcSamples}
\title{Simulate Monte Carlo Samples from a Multivariate Normal Distribution}
\usage{
mcSamples(meanVec, covMat, samples, pops, ..., matrix = FALSE)
}
\arguments{
\item{meanVec}{Numeric vector of means for the variables.}

\item{covMat}{Positive-definite symmetric matrix of the variables.}

\item{samples}{samples}

\item{pops}{populations}

\item{...}{Other variables used in mvnorm function in the mass package.}

\item{matrix}{matrix}
}
\value{
Matrix of maxN multivariate normal samples.
}
\description{
Simulate Monte Carlo Samples from a Multivariate Normal Distribution
}
\keyword{internal}
