% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schott2001.R
\name{Schott2001}
\alias{Schott2001}
\title{Test of Equality of Covariances given by Schott 2001}
\usage{
Schott2001(x, ...)
}
\arguments{
\item{x}{data as a data frame, list of matrices, grouped data frame, or resample object}

\item{...}{other options passed to covTest method}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Schott 2001
}
\examples{
irisSpecies <- unique(iris$Species)

iris_ls <- lapply(irisSpecies, 
    function(x){as.matrix(iris[iris$Species == x, 1:4])}
                 )
                 
names(iris_ls) <- irisSpecies

Chaipitak2013(iris_ls)
}
