% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schott2007_test.R
\name{Schott2007_test}
\alias{Schott2007_test}
\alias{Schott2007_test.data.frame}
\alias{Schott2007_test.grouped_df}
\alias{Schott2007_test.matrix}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Schott2007_test(x, ...)

\method{Schott2007_test}{data.frame}(x, group, ...)

\method{Schott2007_test}{grouped_df}(x, ...)

\method{Schott2007_test}{matrix}(...)
}
\arguments{
\item{x}{tidy data frame}

\item{...}{other}

\item{group}{group}
}
\value{
Test Statistic for Schott 2007
}
\description{
Test of Equality of Covariances given by Schott 2007
}
\examples{
Schott2007_test(iris, group = Species)

}

