% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schott2007_test.R
\name{Schott2007_test}
\alias{Schott2007_test}
\alias{Schott2007_test.data.frame}
\alias{Schott2007_test.grouped_df}
\alias{Schott2007_test.matrix}
\alias{Schott2007_test.resample}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Schott2007_test(x, ...)

\method{Schott2007_test}{data.frame}(x, group, ...)

\method{Schott2007_test}{grouped_df}(x, ...)

\method{Schott2007_test}{resample}(x, ...)

\method{Schott2007_test}{matrix}(...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}

\item{group}{group or population variable}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Schott 2007
}
\details{
Consider the test for the equality of covariance matrices with
\deqn{\text{H}_0: \boldsymbol{\Sigma}_1 = \boldsymbol{\Sigma}_2 = \ldots = \boldsymbol{\Sigma}_k}
and
\deqn{\text{H}_{\text{A}}: \boldsymbol{\Sigma}_1 \neq \boldsymbol{\Sigma}_2 \neq \ldots \neq \boldsymbol{\Sigma}_k,}
where \eqn{\boldsymbol{\Sigma_i}} are the population covariance matrix parameters with \eqn{n_i\textbf{S}_i \sim W_k\left( \boldsymbol{\Sigma}_i, n_i \right)}

\deqn{T_{Sc} = \sum \limits^{k}_{i < j} \frac{ \left( \hat{a}_{2i_{Sc}} + \hat{a}_{2j_{Sc}} - \frac{2}{p}tr \left( S_i S_j \right) \right) ^ 2}{\theta_{Sc}}}
}
\examples{
Schott2007_test(iris, group = Species)

}
\references{
Schott, J. (2007). A test for the equality of covariance matrices when the dimension
is large relative to the sample sizes. Computational Statistics & Data Analysis, 51(12):6535-6542.
}

