% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schott2007_test.R
\name{Schott2007_test}
\alias{Schott2007_test}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Schott2007_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\value{
Test Statistic for Schott 2007
}
\description{
Test of Equality of Covariances given by Schott 2007
}
\examples{
Schott2007_test(mcSamples(c(0,0,0), diag(1, 3), 10, 2, matrix = FALSE, tidy = TRUE), group = population, variables = variable, samples = samples, value = value, tidy = TRUE)

}

