% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{Schott2007_test}
\alias{Schott2007_test}
\title{Test of Equality of Covariances given by Schott 2007}
\usage{
Schott2007_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\description{
Test of Equality of Covariances given by Schott 2007
}

