% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songEquality.R
\name{songEquality}
\alias{songEquality}
\title{Song Conical Variates For Structure of Covariance Tests}
\usage{
songEquality(x)
}
\arguments{
\item{x}{data}

\item{Sigma}{covariance matrix}
}
\description{
Song Conical Variates For Structure of Covariance Tests
}
\keyword{internal}
