% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2007.R
\name{Srivastava2007}
\alias{Srivastava2007}
\title{Test of Equality of Covariances given by Srivastava 2007}
\usage{
Srivastava2007(x, ...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Srivastava 2007
}
\examples{
Srivastava2007(iris, group = Species)

}
\references{
Srivastava, M. S. (2007). Testing the equality of two covariance matrices and
independence of two sub-vectors with fewer observations than the dimension. InInternational
Conference on Advances in InterdisciplinaryStistics and Combinatorics, University of North Carolina
at Greensboro, NC, USA.
}
