% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2007_test.R
\name{Srivastava2007_test}
\alias{Srivastava2007_test}
\alias{Srivastava2007_test.data.frame}
\alias{Srivastava2007_test.default}
\alias{Srivastava2007_test.grouped_df}
\alias{Srivastava2007_test.matrix}
\title{Test of Equality of Covariances given by Srivastava 2007}
\usage{
Srivastava2007_test(data, ...)

\method{Srivastava2007_test}{data.frame}(x, group, ...)

\method{Srivastava2007_test}{grouped_df}(x, ...)

\method{Srivastava2007_test}{matrix}(...)

\method{Srivastava2007_test}{default}(ahat2i, etahat2i)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\value{
Test statistic for Srivastava 2007
}
\description{
Test of Equality of Covariances given by Srivastava 2007
}
\examples{
Srivastava2007_test(iris, group = Species)

}

