% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{Srivastava2007_test}
\alias{Srivastava2007_test}
\title{Test of Equality of Covariances given by Srivastava 2007}
\usage{
Srivastava2007_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\description{
Test of Equality of Covariances given by Srivastava 2007
}

