% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2014_test.R
\name{Srivastava2014_test}
\alias{Srivastava2014_test}
\alias{Srivastava2014_test.data.frame}
\alias{Srivastava2014_test.grouped_df}
\alias{Srivastava2014_test.matrix}
\alias{Srivastava2014_test.resample}
\title{Test of Equality of Covariances given by Srivastava et al. 2014}
\usage{
Srivastava2014_test(x, ...)

\method{Srivastava2014_test}{data.frame}(x, group, ...)

\method{Srivastava2014_test}{grouped_df}(x, ...)

\method{Srivastava2014_test}{resample}(x, ...)

\method{Srivastava2014_test}{matrix}(...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}

\item{group}{group or population variable}
}
\value{
Test statistic of the hypothesis test
}
\description{
Performs 2 and k sample equality of covariance matrix test using Srivastava et al. 2014
}
\examples{
Srivastava2014_test(iris, group = Species)

}
\references{
Srivastava, M. S., Yanagihara, H., and Kubokawa, T. (2014). Tests for covariance
matrices in high dimensionwith less sample size.Journal of Multivariate Analysis, 130:289–309.
}

