% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srivastava2014_test.R
\name{Srivastava2014_test}
\alias{Srivastava2014_test}
\title{Test of Equality of Covariances given by Srivastava 2014}
\usage{
Srivastava2014_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\value{
Test Statistic for Srivastava 2014
}
\description{
Test of Equality of Covariances given by Srivastava 2014
}
\examples{
Srivastava2014_test(mcSamples(c(0,0,0), diag(1, 3), 10, 3), group = population)

}

