% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrivastavaYanagihara2010_test.R
\name{SrivastavaYanagihara2010}
\alias{SrivastavaYanagihara2010}
\title{Hidden Test}
\usage{
SrivastavaYanagihara2010(gammahat_ls, ksihat2_ls)
}
\arguments{
\item{gammahat_ls}{gamma hat}

\item{ksihat2_ls}{ksi hat squared}
}
\description{
Hidden Test
}
