% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrivastavaYanagihara2010_test.R
\name{SrivastavaYanagihara2010_test}
\alias{SrivastavaYanagihara2010_test}
\alias{SrivastavaYanagihara2010_test.data.frame}
\alias{SrivastavaYanagihara2010_test.grouped_df}
\alias{SrivastavaYanagihara2010_test.matrix}
\title{Test of Equality of Covariances given by Srivastava and Yanagihara 2010}
\usage{
SrivastavaYanagihara2010_test(x, ...)

\method{SrivastavaYanagihara2010_test}{data.frame}(x, group, ...)

\method{SrivastavaYanagihara2010_test}{grouped_df}(x, ...)

\method{SrivastavaYanagihara2010_test}{matrix}(...)
}
\arguments{
\item{x}{data as data.frame, grouped_df, resample or matrix object}

\item{...}{other options passed to functions}

\item{group}{group or population variable}
}
\value{
Test statistic for Srivastava and Yanagihara 2010
}
\description{
Test of Equality of Covariances given by Srivastava and Yanagihara 2010
}
\examples{
SrivastavaYanagihara2010_test(iris, group = Species)

}
\references{
Srivastava, M. and Yanagihara, H. (2010). Testing the equality of several covariance matrices with
fewer observation that the dimension. Journal of Multivariate Analysis, 101(6):1319-1329.
}
