% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrivastavaYanagihara2010_test.R
\name{SrivastavaYanagihara2010_test}
\alias{SrivastavaYanagihara2010_test}
\title{Test of Equality of Covariances given by Srivastava and Yanagihara 2010}
\usage{
SrivastavaYanagihara2010_test(data, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{...}{other}
}
\value{
Test statistic for Srivastava and Yanagihara 2010
}
\description{
Test of Equality of Covariances given by Srivastava and Yanagihara 2010
}
\examples{
SrivastavaYanagihara2010_test(mcSamples(rep(0, 100), diag(1, 100), 10, 2), group = population)

}

