% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_functions.R
\name{tau_func}
\alias{tau_func}
\title{Tau for Chaipitak 2013 (helper function)}
\usage{
tau_func(ns)
}
\arguments{
\item{ns}{sample size}
}
\description{
Tau for Chaipitak 2013 (helper function)
}
\keyword{internal}

