% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_sim.R
\name{tests}
\alias{tests}
\title{Title}
\usage{
tests(directory, test_funcs, dimensions)
}
\arguments{
\item{type}{}
}
\description{
Title
}
\examples{
tests(directory = "~/Documents/R/Dissertation/data",
                test_funcs = c("Chaipitak2013_test"),
                dimensions = c(20, 19, 18))
}

